requirements.txt
numpy==1.21.6
torch_scatter==2.1.2
torch_sparse==0.6.18
torch_geometric==2.6.1
scipy==1.10.1
matplotlib==3.7.5
scikit-learn==1.3.2
ogb==1.3.6
pyyaml==6.0.2
deeprobust==0.2.11
seaborn==0.13.2
munkres==1.1.4
cvxpy==1.5.2
pymetis==2023.1.1
ipdb==0.13.13

